[REPLAY] Move file to match with function name
