+++ packages/rum-core/src/index.ts
@@ -30,4 +30,4 @@ export { RumInitConfiguration, RumConfiguration } from './domain/configuration'
 export { DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE } from './domain/rumEventsCollection/action/getActionNameFromElement'
 export { STABLE_ATTRIBUTES } from './domain/rumEventsCollection/action/getSelectorFromElement'
 export * from './browser/htmlDomUtils'
-export { getSessionReplayUrl } from './domain/getDatadogUrl'
+export { getSessionReplayUrl } from './domain/getSessionReplayUrl'
